aidevs3/w4/s04e01/requirements.txt
openai
ollama
PyYAML
requests
wget
